* test/win32/testlibwin32.cpp - Include emulation header - Rename _chdir back to chdir, since emu.h #defines the new to the old
* lib/win32/emu.cpp
- Moved '#include <process.h>' to sit next to unistd.h
- Improved comment formatting
- Renamed tTimer to Timer_t

* lib/backupclient/BackupClientRestore.cpp
* lib/backupclient/BackupStoreFileDiff.cpp
* lib/backupclient/BackupClientFileAttributes.cpp
* lib/backupclient/BackupStoreFileEncodeStream.cpp
* lib/backupclient/BackupStoreDirectory.h
* lib/backupclient/BackupStoreFileCombine.cpp
* lib/backupclient/BackupStoreFile.cpp
* lib/backupclient/BackupStoreFilename.cpp
* lib/backupclient/BackupStoreFilenameClear.cpp
* lib/backupclient/BackupStoreFileCmbDiff.cpp
* lib/backupclient/BackupStoreDirectory.cpp
* lib/backupclient/BackupStoreFileCmbIdx.cpp
* lib/backupclient/BackupStoreFile.h
* lib/backupclient/BackupStoreFileRevDiff.cpp
* lib/server/Protocol.cpp
* lib/server/SocketStream.cpp
* lib/server/Protocol.h
* lib/server/SocketStreamTLS.cpp
* lib/common/ReadGatherStream.cpp
* lib/common/IOStreamGetLine.cpp
* lib/common/FdGetLine.cpp
* lib/common/MemBlockStream.h
* lib/common/WaitForEvent.cpp
* lib/common/CollectInBufferStream.cpp
* lib/common/StreamableMemBlock.cpp
* lib/common/Utils.cpp
* lib/common/MemBlockStream.cpp
* lib/common/CollectInBufferStream.h
* lib/common/ExcludeList.h
* lib/common/BoxTimeToUnix.h
* bin/bbackupctl/bbackupctl.cpp
* bin/bbackupd/BackupClientDirectoryRecord.cpp
* bin/bbackupd/BackupDaemon.cpp
- Undo type changes and type casts
- Undo formatting changes to reduce white space noise

* lib/common/Test.h
* lib/common/Guards.h
- Remove #included "Box.h" from header file as requested by Ben

+++ lib/backupclient/BackupClientFileAttributes.cpp
@@ -632,7 +632,7 @@ void BackupClientFileAttributes::WriteAttributes(const char *Filename) const
 		}
 #endif
 
-		xattrOffset += (int)std::strlen(reinterpret_cast<char*>(pattr+1))+1;
+		xattrOffset += std::strlen(reinterpret_cast<char*>(pattr+1))+1;
 	}
 	
 	// If working as root, set user IDs
@@ -986,7 +986,7 @@ uint64_t BackupClientFileAttributes::GenerateAttributeHash(struct stat &st, cons
 	MD5Digest digest;
 	digest.Add(&hashData, sizeof(hashData));
 	digest.Add(xattr.GetBuffer(), xattr.GetSize());
-	digest.Add(leafname.c_str(), (int)leafname.size());
+	digest.Add(leafname.c_str(), leafname.size());
 	digest.Add(sAttributeHashSecret, sAttributeHashSecretLength);
 	digest.Finish();
